+++ storage/volume.go
@@ -40,6 +40,7 @@ type VolumeConfig struct {
 	QoS                       string                 `json:"qos,omitempty"`
 	QoSType                   string                 `json:"type,omitempty"`
 	ServiceLevel              string                 `json:"serviceLevel,omitempty"`
+	Network                   string                 `json:"network,omitempty"`
 	ImportOriginalName        string                 `json:"importOriginalName,omitempty"`
 	ImportBackendUUID         string                 `json:"importBackendUUID,omitempty"`
 	ImportNotManaged          bool                   `json:"importNotManaged,omitempty"`